+++ app/Http/Controllers/PreferenceController.php
@@ -4,7 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests;
 use App\Preference;
-use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Request;
 
 class PreferenceController extends Controller
 {
@@ -15,8 +15,17 @@ class PreferenceController extends Controller
      */
     public function index()
     {
-        $all = Preference::all();
-        return response()->json($all);
+        $var_token = Request::header('api_token');
+        $user_token = User::where('api_token',$var_token)->first();
+
+        if($user_token != null) {
+            $all = Preference::all();
+            return response()->json($all);
+        } else { return response()->json(array(
+                        'code'      =>  404,
+                        'message'   =>  'Usuário não autenticado'
+                        ), 404);
+                }
     }
 
     /**
@@ -27,9 +36,18 @@ class PreferenceController extends Controller
      */
     public function store(Request $request)
     {
-        $preference = $request->all();
-        $preference = Preference::create($preference);
-        return response()->json($preference);
+        $var_token = Request::header('api_token');
+        $user_token = User::where('api_token',$var_token)->first();
+
+        if($user_token != null) {
+            $preference = $request->all();
+            $preference = Preference::create($preference);
+            return response()->json($preference);
+        } else { return response()->json(array(
+                        'code'      =>  404,
+                        'message'   =>  'Usuário não autenticado'
+                        ), 404);
+                }
     }
 
     /**
@@ -40,8 +58,17 @@ class PreferenceController extends Controller
      */
     public function show($id)
     {
-        $all = Preference::findOrFail($id);
-        return response()->json($all);
+        $var_token = Request::header('api_token');
+        $user_token = User::where('api_token',$var_token)->first();
+
+        if($user_token != null) {
+            $all = Preference::findOrFail($id);
+            return response()->json($all);
+        } else { return response()->json(array(
+                        'code'      =>  404,
+                        'message'   =>  'Usuário não autenticado'
+                        ), 404);
+                }
     }
 
     /**
@@ -53,10 +80,19 @@ class PreferenceController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $preference = Preference::findOrFail($id);
-        $input = $request->all();
-        $preference->fill($input)->save();
-        return response()->json($preference);
+        $var_token = Request::header('api_token');
+        $user_token = User::where('api_token',$var_token)->first();
+
+        if($user_token != null) {
+            $preference = Preference::findOrFail($id);
+            $input = $request->all();
+            $preference->fill($input)->save();
+            return response()->json($preference);
+        } else { return response()->json(array(
+                        'code'      =>  404,
+                        'message'   =>  'Usuário não autenticado'
+                        ), 404);
+                }
     }
 
     /**
@@ -67,7 +103,16 @@ class PreferenceController extends Controller
      */
     public function destroy($id)
     {
-        $preference = Preference::findOrFail($id);
-        $preference->delete();
+        $var_token = Request::header('api_token');
+        $user_token = User::where('api_token',$var_token)->first();
+
+        if($user_token != null) {
+            $preference = Preference::findOrFail($id);
+            $preference->delete();
+        } else { return response()->json(array(
+                        'code'      =>  404,
+                        'message'   =>  'Usuário não autenticado'
+                        ), 404);
+                }
     }
 }
