+++ app/Http/Controllers/ResidenceController.php
@@ -8,8 +8,8 @@ use App\Preference;
 use App\Residence;
 use App\ResidenceImage;
 use App\User;
-use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Request;
 use Intervention\Image\Facades\Image;
 
 class ResidenceController extends Controller
@@ -21,7 +21,7 @@ class ResidenceController extends Controller
      */
     public function index()
     {
-        $var_token = $_GET['api_token'];
+        $var_token = Request::header('api_token');
         $user_token = User::where('api_token',$var_token)->first();
 
         if($user_token != null) {
@@ -35,32 +35,65 @@ class ResidenceController extends Controller
     }
 
     /**
+    * Display a listing of the resource.
+    *
+    * @return \Illuminate\Http\Response
+    */
+    public function residences($idUser,$id,$qtReg){
+       $var_token = Request::header('api_token');
+       $user_token = User::where('api_token',$var_token)->first();
+        if($user_token != null) {
+           $residence = Residence::with('user')
+                        ->where('idUser',$idUser)
+                        ->limit($qtReg)
+                        ->offset($id-1)
+                        ->get();
+           return response()->json($residence);
+      }else
+       { return response()->json(array(
+                           'code'      =>  404,
+                           'message'   =>  'Usuário não autenticado'
+                           ), 404);
+       }
+    }    
+
+    /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
      */
     public function near($idUser){
-        $user = User::with('location')->findOrFail($idUser);
-  
-        $distanceW = DB::table('location')
-                     ->select(DB::raw('idLocation,substr((6371 *
-                                                acos(
-                                                    cos(radians('.$user->location['latitude'].')) *
-                                                    cos(radians(latitude)) *
-                                                    cos(radians('.$user->location['longitude'].') - radians(longitude)) +
-                                                    sin(radians('.$user->location['latitude'].')) *
-                                                    sin(radians(latitude))
-                                                )),1,4) AS distanceW'))
-                     ->having('distanceW', '<', $user->distance)
-                     ->get();
-        //return count($distanceW);
-      //arrumar a distance.
-      /*  for ($i = 0; $i <= count($distanceW); ++$i){
-            $distanceTeste = $distanceW[$i]['idLocation'];
-        }
-        echo $distanceTeste;
-        *///$residence = Residence::with('location')->where('idLocation','in',$locations)->get();
-        //return response()->json($residence);
+        $var_token = Request::header('api_token');
+        $user_token = User::where('api_token',$var_token)->first();
+
+        if($user_token != null) {
+            $user = User::with('location')->findOrFail($idUser);
+            $distanceW = DB::table('location')
+                         ->select(DB::raw('idLocation,substr((6371 *
+                                                    acos(
+                                                        cos(radians('.$user->location['latitude'].')) *
+                                                        cos(radians(latitude)) *
+                                                        cos(radians('.$user->location['longitude'].') - radians(longitude)) +
+                                                        sin(radians('.$user->location['latitude'].')) *
+                                                        sin(radians(latitude))
+                                                    )),1,4) AS distanceW'))
+                         ->where('idLocation','<>',$user->location['idLocation'])
+                         ->having('distanceW', '<', $user->distance)
+                         ->get();
+
+            $distanceTeste = $distanceW[0]->idLocation;             
+            for ($i = 1; $i <= count($distanceW)-1; ++$i){
+                $distanceTeste = $distanceTeste.','.$distanceW[$i]->idLocation;        
+            }
+            $residences = Residence::with('location')
+                         ->where('idLocation', $distanceTeste)->get();
+            
+            return response()->json($residences);
+        } else { return response()->json(array(
+                        'code'      =>  404,
+                        'message'   =>  'Usuário não autenticado'
+                        ), 404);
+                }
     }
 
     /**
@@ -70,11 +103,20 @@ class ResidenceController extends Controller
      */
     public function limit($id,$qtReg)
     {
-        $sql = Residence::with('location','user','preference')
-                     ->limit($qtReg)
-                     ->offset($id-1)
-                     ->get();    
-        return response()->json($sql);
+       $var_token = Request::header('api_token');
+       $user_token = User::where('api_token',$var_token)->first();
+        if($user_token != null) {
+            $sql = Residence::with('location','user','preference')
+                      ->limit($qtReg)
+                      ->offset($id-1)
+                      ->get();    
+            return response()->json($sql);
+      }else
+       { return response()->json(array(
+                           'code'      =>  404,
+                           'message'   =>  'Usuário não autenticado'
+                           ), 404);
+       }
     }
 
         /**
@@ -84,10 +126,19 @@ class ResidenceController extends Controller
      */
     public function residenceImages($idResidence)
     {
-        $residenceImage = Residence::with('residenceImages')
-            ->where('idResidence', '=', $idResidence)
-            ->get();    
-        return response()->json($residenceImage);
+        $var_token = Request::header('api_token');
+        $user_token = User::where('api_token',$var_token)->first();
+
+        if($user_token != null) {
+            $residenceImage = Residence::with('residenceImages')
+                ->where('idResidence', '=', $idResidence)
+                ->get();    
+            return response()->json($residenceImage);
+        } else { return response()->json(array(
+                        'code'      =>  404,
+                        'message'   =>  'Usuário não autenticado'
+                        ), 404);
+                }
     }
 
        /**
@@ -97,17 +148,26 @@ class ResidenceController extends Controller
      */
     public function eager($idResidence)
     {
-        $eager = Residence::with(['matches' => function($query){
-            $query->orderBy('dateTime','desc');
-        },                        'residenceImages' => function($query2){
-            $query2->orderBy('orderImage','desc');                        
-        },
-                                  'location.city',
-                                  'preference',
-                                  'user',
-                                  'matches.user'])
-        ->findOrFail($idResidence);
-        return response()->json($eager);
+        $var_token = Request::header('api_token');
+        $user_token = User::where('api_token',$var_token)->first();
+
+        if($user_token != null) {
+             $eager = Residence::with(['matches' => function($query){
+                $query->orderBy('dateTime','desc');
+            },                        'residenceImages' => function($query2){
+                $query2->orderBy('orderImage','desc');                        
+            },
+            'location.city',
+            'preference',
+            'user',
+            'matches.user'])
+             ->findOrFail($idResidence);
+             return response()->json($eager);
+        } else { return response()->json(array(
+                        'code'      =>  404,
+                        'message'   =>  'Usuário não autenticado'
+                        ), 404);
+                }
     }
 
     /**
@@ -115,15 +175,23 @@ class ResidenceController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    //IMAGEM PRINCIPAL DA RESIDENCIA
     public function top($idResidence)
     {
-        $residenceImage = Residence::with(['residenceImages' => function($query){
-            $query->orderBy('orderImage','desc')->first();
-        }])
-            ->where('idResidence', '=', $idResidence)
-            ->first(); 
-        return response()->json($residenceImage);
+        $var_token = Request::header('api_token');
+        $user_token = User::where('api_token',$var_token)->first();
+
+        if($user_token != null) {
+            $residenceImage = Residence::with(['residenceImages' => function($query){
+                $query->orderBy('orderImage','desc')->first();
+            }])
+                ->where('idResidence', '=', $idResidence)
+                ->first(); 
+            return response()->json($residenceImage);
+        } else { return response()->json(array(
+                        'code'      =>  404,
+                        'message'   =>  'Usuário não autenticado'
+                        ), 404);
+                }
     }
 
     /**
@@ -134,69 +202,50 @@ class ResidenceController extends Controller
      */
     public function store(Request $request)
     {   
-        try
-        {
-            DB::beginTransaction();  
+        $var_token = Request::header('api_token');
+        $user_token = User::where('api_token',$var_token)->first();
 
-            $residence = $request->get('jsonObject');
-            //retorna array, se tirar o true do json_decode vai retornar um objeto e vai dar erro em tudo.
-            $residence = json_decode($residence,true);
+        if($user_token != null) {
+            try
+            {
+                DB::beginTransaction();  
 
-            //Cria a localização desta residencia.
-            $location = $residence['location'];
-            $location = Location::create($location);
+                $residence = Request::header('jsonObject');
+                //retorna array, se tirar o true do json_decode vai retornar um objeto e vai dar erro em tudo.
+                $residence = json_decode($residence,true);
 
-            //Cria as preferencias da residencia.
-            $preference = $user['preference'];
-            $preference = Preference::create($preference);
-            
-            $residence = Residence::create([
-                'idLocation' => $location['idLocation'],
-                'idUser' => $residence['idUser'],
-                'idPreference' => $preference['idPreference'],
-                'title' => $residence['title'],
-                'description' => $residence['description'],
-                'address' => $residence['address'],
-                'observation' => $residence['observation'],
-                'rent' => $residence['rent'],
-                'tries' => 0,   
-                'active' => true,
-            ]);
+                //Cria a localização desta residencia.
+                $location = $residence['location'];
+                $location = Location::create($location);
 
-            //Adiciona as imagens da residencia.
-           /* $residenceImages = $request->residenceImages;
-            foreach($residenceImages as $residenceImage){
-                $file = $residenceImage['path'];
-                $size = getimagesize($file); 
-                list($type, $file) = explode(';', $file);
-                list(,$file) = explode(',', $file);
-                $file = base64_decode($file);
-
-                $name_image = "residenceImage_".time().".png";
-
-                $path_high = public_path()."/img/residenceImage/high/".$name_image;
-                $path_medium = public_path()."/img/residenceImage/medium/".$name_image;
-                $path_low = public_path()."/img/residenceImage/low/".$name_image;
-
-                Image::make($file)->save($path_high);
-                Image::make($file)->resize($size[0]/3, $size[1]/3)->save($path_medium);
-                Image::make($file)->resize($size[0]/5, $size[1]/5)->save($path_low);
-
-                $residenceImageCreate = ResidenceImage::create([
-                    'idResidence' => $residence['idResidence'],
-                    'path' => $name_image,
-                    'resource' => $residenceImage['resource'],
-                    'orderImage' => $residenceImage['orderImage'],
-                    'tries' => 0,
+                //Cria as preferencias da residencia.
+                $preference = $user['preference'];
+                $preference = Preference::create($preference);
+                
+                $residence = Residence::create([
+                    'idLocation' => $location['idLocation'],
+                    'idUser' => $residence['idUser'],
+                    'idPreference' => $preference['idPreference'],
+                    'title' => $residence['title'],
+                    'description' => $residence['description'],
+                    'address' => $residence['address'],
+                    'observation' => $residence['observation'],
+                    'rent' => $residence['rent'],
+                    'tries' => 0,   
                     'active' => true,
-                    ]);   
-            }*/
-            DB::Commit();
-            return response()->json($residence);
-        } catch (Exception $e){
-            DB::Rollback();
-            return response()->$e;
-        }      
+                ]);
+                
+                DB::Commit();
+                return response()->json($residence);
+            } catch (Exception $e){
+                DB::Rollback();
+                return response()->$e;
+            }      
+        } else { return response()->json(array(
+                        'code'      =>  404,
+                        'message'   =>  'Usuário não autenticado'
+                        ), 404);
+                }
     }
 
     /**
@@ -207,8 +256,17 @@ class ResidenceController extends Controller
      */
     public function show($id)
     {
-        $all = Residence::with('location','user','preference')->findOrFail($id);
-        return response()->json($all);
+        $var_token = Request::header('api_token');
+        $user_token = User::where('api_token',$var_token)->first();
+
+        if($user_token != null) {
+            $all = Residence::with('location','user','preference')->findOrFail($id);
+            return response()->json($all);
+        } else { return response()->json(array(
+                        'code'      =>  404,
+                        'message'   =>  'Usuário não autenticado'
+                        ), 404);
+                }
     }
 
     /**
@@ -220,10 +278,19 @@ class ResidenceController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $residence = Residence::findOrFail($id);
-        $input = $request->all();
-        $residence->fill($input)->save();
-        return response()->json($residence);
+        $var_token = Request::header('api_token');
+        $user_token = User::where('api_token',$var_token)->first();
+
+        if($user_token != null) {
+            $residence = Residence::findOrFail($id);
+            $input = $request->all();
+            $residence->fill($input)->save();
+            return response()->json($residence);
+        } else { return response()->json(array(
+                        'code'      =>  404,
+                        'message'   =>  'Usuário não autenticado'
+                        ), 404);
+                }
     }
 
     /**
@@ -234,7 +301,16 @@ class ResidenceController extends Controller
      */
     public function destroy($id)
     {
-        $residence = Residence::findOrFail($id);
-        $residence->delete();
+        $var_token = Request::header('api_token');
+        $user_token = User::where('api_token',$var_token)->first();
+
+        if($user_token != null) {
+            $residence = Residence::findOrFail($id);
+            $residence->delete();
+        } else { return response()->json(array(
+                        'code'      =>  404,
+                        'message'   =>  'Usuário não autenticado'
+                        ), 404);
+                }
     }
 }
